Moved safeguards/preprocessing.go into its own subdirectory under safeguards package.
