+++ pkg/safeguards/preprocessing/preprocessing_helpers_test.go
@@ -1,4 +1,4 @@
-package safeguards
+package preprocessing
 
 import (
 	"os"
